requirements.txt
streamlit==1.32.0
requests==2.31.0
pandas==2.2.1
plotly==5.18.0
python-pptx==0.6.23
python-dotenv==1.0.1
kaleido==0.2.1
wordcloud==1.9.3
matplotlib==3.8.3
numpy==1.26.4
Pillow==11.1.0